+++ python/pyspark/pandas/tests/test_generic_functions.py
@@ -22,129 +22,6 @@ from pyspark.testing.pandasutils import PandasOnSparkTestCase, TestUtils
 
 
 class GenericFunctionsTestsMixin:
-    def test_interpolate_error(self):
-        psdf = ps.range(10)
-
-        with self.assertRaisesRegex(
-            NotImplementedError, "interpolate currently works only for method='linear'"
-        ):
-            psdf.interpolate(method="quadratic")
-
-        with self.assertRaisesRegex(
-            NotImplementedError, "interpolate currently works only for method='linear'"
-        ):
-            psdf.id.interpolate(method="quadratic")
-
-        with self.assertRaisesRegex(ValueError, "limit must be > 0"):
-            psdf.interpolate(limit=0)
-
-        with self.assertRaisesRegex(ValueError, "limit must be > 0"):
-            psdf.id.interpolate(limit=0)
-
-        with self.assertRaisesRegex(ValueError, "invalid limit_direction"):
-            psdf.interpolate(limit_direction="jump")
-
-        with self.assertRaisesRegex(ValueError, "invalid limit_direction"):
-            psdf.id.interpolate(limit_direction="jump")
-
-        with self.assertRaisesRegex(ValueError, "invalid limit_area"):
-            psdf.interpolate(limit_area="jump")
-
-        with self.assertRaisesRegex(ValueError, "invalid limit_area"):
-            psdf.id.interpolate(limit_area="jump")
-
-    def _test_interpolate(self, pobj):
-        psobj = ps.from_pandas(pobj)
-        self.assert_eq(psobj.interpolate(), pobj.interpolate())
-        for limit in range(1, 5):
-            for limit_direction in [None, "forward", "backward", "both"]:
-                for limit_area in [None, "inside", "outside"]:
-                    self.assert_eq(
-                        psobj.interpolate(
-                            limit=limit, limit_direction=limit_direction, limit_area=limit_area
-                        ),
-                        pobj.interpolate(
-                            limit=limit, limit_direction=limit_direction, limit_area=limit_area
-                        ),
-                    )
-
-    def test_interpolate(self):
-        pser = pd.Series(
-            [
-                1,
-                np.nan,
-                3,
-            ],
-            name="a",
-        )
-        self._test_interpolate(pser)
-
-        pser = pd.Series(
-            [
-                np.nan,
-                np.nan,
-                np.nan,
-            ],
-            name="a",
-        )
-        self._test_interpolate(pser)
-
-        pser = pd.Series(
-            [
-                np.nan,
-                np.nan,
-                np.nan,
-                0,
-                1,
-                np.nan,
-                np.nan,
-                np.nan,
-                np.nan,
-                3,
-                np.nan,
-                np.nan,
-                np.nan,
-            ],
-            name="a",
-        )
-        self._test_interpolate(pser)
-
-        pdf = pd.DataFrame(
-            [
-                (1, 0.0, np.nan),
-                (2, np.nan, 2.0),
-                (3, 2.0, 3.0),
-                (4, np.nan, 4.0),
-                (5, np.nan, 1.0),
-            ],
-            columns=list("abc"),
-        )
-        self._test_interpolate(pdf)
-
-        pdf = pd.DataFrame(
-            [
-                (0.0, np.nan, -1.0, 1.0, np.nan),
-                (np.nan, 2.0, np.nan, np.nan, np.nan),
-                (2.0, 3.0, np.nan, 9.0, np.nan),
-                (np.nan, 4.0, -4.0, 16.0, np.nan),
-                (np.nan, 1.0, np.nan, 7.0, np.nan),
-            ],
-            columns=list("abcde"),
-        )
-        self._test_interpolate(pdf)
-
-        pdf = pd.DataFrame(
-            [
-                (0.0, np.nan, -1.0, False, np.nan),
-                (np.nan, 2.0, np.nan, True, np.nan),
-                (2.0, 3.0, np.nan, True, np.nan),
-                (np.nan, 4.0, -4.0, False, np.nan),
-                (np.nan, 1.0, np.nan, True, np.nan),
-            ],
-            columns=list("abcde"),
-        )
-        self._test_interpolate(pdf)
-
     def _test_stat_functions(self, stat_func):
         pdf = pd.DataFrame({"a": [np.nan, np.nan, np.nan], "b": [1, np.nan, 2], "c": [1, 2, 3]})
         psdf = ps.from_pandas(pdf)
